Rename Home Owners composable
